Add files via upload
This is an R Notebook evaluating the effects of breathing techniques and aerobic exercise training on the physical fitness of pediatric cystic fibrosis patients. The scientific question being answered is "How does an active cycle of breathing techniques and aerobic exercise training affect the physical fitness, as measured by muscle endurance and speed, of young children between the ages of 5-13 with cystic fibrosis?" Furthermore, the hypothesis of this analysis is "If clinically stable pediatric cystic fibrosis patients undergo an active cycle of breathing techniques in addition to aerobic exercise training, then their thoracic mobility and physical fitness parameters such as muscle endurance and speed will increase." Data was obtained from the article titled "Effects of chest physiotherapy and aerobic exercise training on physical fitness in young children with cystic fibrosis." In order to address the question, a paired sample t-test and Deep Neural Network will be done. Finally, histograms will be generated to visualize the differences between before training and after training for the sit-up test and shuttle run test.